fix: modify style - Add cursor pointer - Add styles nth-child to table - Eliminate folder src of blocks
