+++ active-customer/src/blocks/table-data/index.less
@@ -1,0 +1,7 @@
+.rowTab {
+  cursor: pointer;
+}
+
+.rowTab:nth-child(even) {
+  background-color: #f2f2f2;
+}
