+++ active-customer/src/blocks/table-data/interfaces/dataTable.interface.tsx
