+++ active-customer/src/locales/es-ES.ts
@@ -3,6 +3,6 @@ export default {
   'tableContainer.Table-col2':'Nombre de cliente',
   'tableContainer.Table-col3':'Productos activos',
   'tableContainer.Table-col4':'Solicitudes activas',
-  'tableContainer.Table-col5':'información',
+  'tableContainer.Table-col5':'Información',
   'tableContainer.Search-title':'Búsqueda por: número de cuenta o nombre del cliente'  
 };
